requirements.txt
Django==3.1.2
django-countries==6.1.3
djangorestframework==3.12.1
behave==1.2.6
psycopg2-binary>=2.8
